types added in correct way into vuex
