+++ proto/front/src/store/modules/modules.types.ts
@@ -1,0 +1,1 @@
+export type TUid = string;
