+++ proto/front/src/store/modules/profiles/profiles.ts
@@ -1,0 +1,15 @@
+import { Module } from 'vuex';
+import { IRootState } from '../../types';
+import { IState } from './profiles.types';
+
+const profiles: Module<IState, IRootState> = {
+  state: {
+    profiles: {},
+  },
+  mutations: {
+  },
+  actions: {
+  },
+};
+
+export default profiles;
